requirements.txt
google-generativeai
schedule
requests